requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioredis==2.0.1
aiosignal==1.3.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpraw==7.8.0
asyncprawcore==2.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
confluent-kafka==2.6.1
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
motor==3.6.0
multidict==6.1.0
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
uvicorn==0.32.1
websocket-client==1.8.0
yarl==1.18.3
